at least the test is going through now, found the bug with the stack

+++ stack.h
@@ -6,7 +6,7 @@
 static const uint16_t START_ADDR = 0x200; // memory for chip8 rom always gets loaded and starts here 
 
 // stack with 16 levels
-uint8_t stack[16];
+uint16_t stack[16];
 // 8-bit stack pointer
 uint8_t sp;
